% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{choose_em_engine}
\alias{choose_em_engine}
\title{Select EM Algorithm Engine Based on System Resources}
\usage{
choose_em_engine()
}
\value{
A function: either \code{run_em_haplo} or \code{run_em_haplo_multicore}.
}
\description{
Chooses between single-threaded or parallel EM based on available CPU cores.
}
