% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregation - Copy.R
\name{compute_hla_segregation2}
\alias{compute_hla_segregation2}
\title{For each family, identify parental haplotypes, infer inheritance pattern per child,
and reconstruct child haplotypes directly using inferred segregation (e.g. AC, BD).}
\usage{
compute_hla_segregation2(hped, collapse = "~")
}
\arguments{
\item{hped}{A tibble with HLA columns and FAMILY_ID, Family_Member labels.}

\item{collapse}{Separator between genes in collapsed haplotype (default = "~").}
}
\value{
A tibble with FAMILY_ID, Family_Member, Segregation, Haplotype
}
\description{
For each family, identify parental haplotypes, infer inheritance pattern per child,
and reconstruct child haplotypes directly using inferred segregation (e.g. AC, BD).
}
