% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{decode_classical_mac}
\alias{decode_classical_mac}
\title{Decode MAC-Encoded Alleles in Classical HLA Columns}
\usage{
decode_classical_mac(
  df,
  loci = c("A", "B", "C", "DRB1", "DRB3", "DRB4", "DRB5", "DQA1", "DQB1", "DPA1", "DPB1")
)
}
\arguments{
\item{df}{A data frame with HLA allele columns (e.g. A_1, B_2, etc).}

\item{loci}{Optional set of classical genes to decode.}
}
\value{
Data frame with decoded MAC values.
}
\description{
Uses immunotation to decode multi-allele codes (MAC) in classical HLA loci.
}
