% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_algorithm.R
\name{em_algorithm}
\alias{em_algorithm}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{EM Algorithm
}\if{html}{\out{</div>}}}
\usage{
em_algorithm(df_raw, collapse = ", ")
}
\arguments{
\item{df_raw}{A tibble containing family HLA typing data.
Used to find if percentage match is conclusive.}

\item{collapse}{Character string used to separate each alleles.}
}
\value{
A tibble in a long format with index, haplotype strings,
em_probability, numbers of common and different alleles,
percentage match, exact match.
}
\description{
Creates haplotype strings from probability via EM algorithm.
}
\details{
Creates haplotype strings for all possibilities and compared with segregation
analysis to determine best matches.

Segregation analysis is the truth (determines haplotype strings)

Packages (Haplo.stats, dplyr, furrr, future) must be installed using
install.packages() and load in with library().
}
