% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{enumerate_diplotypes}
\alias{enumerate_diplotypes}
\title{Exhaustive Diplotype Enumeration}
\usage{
enumerate_diplotypes(geno_row)
}
\arguments{
\item{geno_row}{A single row from collapsed genotype table.}
}
\value{
A list of diplotypes with \code{hap1} and \code{hap2}.
}
\description{
Returns all possible phased haplotype pairs consistent with a multilocus genotype row.
}
