% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{extract_loci}
\alias{extract_loci}
\title{Extract HLA Locus Names from Genotype Table}
\usage{
extract_loci(df)
}
\arguments{
\item{df}{A data frame with standardized column names.}
}
\value{
A character vector of distinct locus names (e.g. A, B, DRB1).
}
\description{
Identifies unique locus names based on *_1 and *_2 suffixes.
}
