% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{fast_diplotype_generator}
\alias{fast_diplotype_generator}
\title{Fast Diplotype Generator (Non-Exhaustive)}
\usage{
fast_diplotype_generator(geno_row)
}
\arguments{
\item{geno_row}{A single row of multilocus genotype.}
}
\value{
A list of two diplotypes: hap1/hap2 and its reverse.
}
\description{
Infers exactly two phased diplotypes per individual by resolving alleles in input order.
}
