% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segregation.R
\name{hla_analysis}
\alias{hla_analysis}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{HLA Haplotype Analysis (Segregation Analysis)
}\if{html}{\out{</div>}}}
\usage{
hla_analysis(hped, collapse = "~", verbose = TRUE)
}
\arguments{
\item{hped}{A tibble containing family HLA typing data. Must include family
columns ("FAMILY_ID" and "Family_Member") and HLA loci should be in pairs
(A_1, A_2, B_1, B_2 etc.).}

\item{collapse}{Character string used to separate each alleles in haplotype
strings (default = "~").}

\item{verbose}{Logical. If 'TRUE' prints out progress message during analysis.}
}
\value{
A tibble in a long format with FAMILY_ID, Family_Member
(children only), Haplotype strings.
}
\description{
Performs HLA haplotype phasing for segregation analysis.
}
\details{
Create haplotype strings (A, B, C, D) for each child per family to determine
inheritance patterns.

More than 3 children must be present for single parent families (for inferred).

Packages (dplyr, purrr and janitor) must be installed using
install.packages() and load in with library().
}
