% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_typing_data.R
\name{load_typing_data}
\alias{load_typing_data}
\title{Load HLA Typing Data from CSV or TSV Using readr}
\usage{
load_typing_data(filepath, ...)
}
\arguments{
\item{filepath}{Path to the data file (.csv or .tsv).}

\item{...}{Additional arguments passed to the readr reading functions.}
}
\value{
A tibble containing the HLA typing data.
}
\description{
Reads an HLA typing data file in CSV or TSV format and returns a tibble.
}
