% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.R, R/global.R
\name{plot_hla_allele_count}
\alias{plot_hla_allele_count}
\title{Plot Count of Unique Alleles per HLA Gene}
\usage{
plot_hla_allele_count(hped)

plot_hla_allele_count(hped)
}
\arguments{
\item{hped}{A data frame or tibble with HLA allele columns.}

\item{freq_data}{A data frame with gene and allele columns.}

\item{min_count}{Minimum unique allele count to include gene (default = 0).}

\item{quiet}{Logical. If TRUE, suppresses status messages.}
}
\value{
A ggplot object that can be further customized or printed.
}
\description{
Visualizes the number of unique observed alleles for each gene.

Visualizes the number of unique observed alleles for each gene.
}
