% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_print_results.R
\name{print.HLAhaploTools}
\alias{print.HLAhaploTools}
\title{Print summary of HLAhaploTools result}
\usage{
\method{print}{HLAhaploTools}(x, ...)
}
\arguments{
\item{x}{Output list returned from \code{HLAhaploTools()}}

\item{...}{Ignored}
}
\description{
Custom print method for objects returned by \code{HLAhaploTools()}.
Displays a summary of key results including allele frequencies,
haplotype inference, family segregation, and deletion checks.
}
