% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{run_em_haplo}
\alias{run_em_haplo}
\title{EM Algorithm for HLA Haplotype Frequency Estimation (Serial)}
\usage{
run_em_haplo(all_diplotypes, epsilon = 1e-05, max_iter = 100)
}
\arguments{
\item{all_diplotypes}{A list of diplotype lists (hap1/hap2) for each individual.}

\item{epsilon}{Numeric. Convergence threshold (default = 1e-5).}

\item{max_iter}{Integer. Maximum EM iterations (default = 100).}
}
\value{
A named numeric vector of estimated haplotype frequencies (sorted descending).
}
\description{
Performs single‐threaded Expectation–Maximization to learn haplotype frequencies
from candidate diplotypes.
}
