% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{run_em_haplo_multicore}
\alias{run_em_haplo_multicore}
\title{Parallel EM Algorithm for HLA Haplotype Frequencies}
\usage{
run_em_haplo_multicore(all_diplotypes, epsilon = 1e-05, max_iter = 100)
}
\arguments{
\item{all_diplotypes}{A list of diplotype lists (hap1/hap2) for each individual.}

\item{epsilon}{Numeric. Convergence threshold (default = 1e-5).}

\item{max_iter}{Integer. Maximum EM iterations (default = 100).}
}
\description{
Accelerates the EM E-step over individuals using future.apply parallelism.
}
