% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{standardize_colnames}
\alias{standardize_colnames}
\title{Standardize HLA Genotype Column Names}
\usage{
standardize_colnames(df)
}
\arguments{
\item{df}{A raw HLA genotype data frame.}
}
\value{
A cleaned data frame with standardized column names.
}
\description{
Converts genotype column names (e.g. A.1) to A_1 format and applies consistent lowercase styling.
}
