% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_deleted_alleles.R
\name{check_deleted_alleles}
\alias{check_deleted_alleles}
\title{Check HLA dataframe for deleted alleles}
\usage{
check_deleted_alleles(
  df_hla,
  deleted_file_path =
    "https://ftp.ebi.ac.uk/pub/databases/ipd/imgt/hla/Deleted_alleles.txt",
  quiet = FALSE
)
}
\arguments{
\item{df_hla}{A dataframe containing HLA allele typing data.}

\item{deleted_file_path}{Path or URL to the Deleted_alleles.txt file
(default: IMGT/HLA database URL).}

\item{quiet}{Logical. If TRUE, suppresses status messages.}
}
\value{
A dataframe containing information about deleted alleles found in the input.
}
\description{
Analyzes a dataframe containing HLA allele designations, checking for
deleted alleles according to the IMGT/HLA database. Downloads and parses the official
deleted allele list, then matches any occurrences in the user's data.
}
\examples{
\dontrun{
results <- check_deleted_alleles(df_hla)
}
}
\keyword{internal}
