% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregation.R
\name{compute_HLA_segregation}
\alias{compute_HLA_segregation}
\title{Compute Sorted and Cleaned HLA Haplotypes (Verbose)}
\usage{
compute_HLA_segregation(hped, collapse = "~")
}
\arguments{
\item{hped}{A data frame with paired HLA columns, including FAMILY_ID and Family_Member.}

\item{collapse}{Separator between genes in collapsed haplotype (default = "~").}
}
\value{
A tibble with FAMILY_ID, Family_Member, haplotype_id, haplotype, and per-locus values.
}
\description{
Extracts haplotypes by sorting allele pairs, decoding multi-alleles,
excluding empty/nulls, and collapsing by gene group within families.
}
