% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{infer_haplotypes}
\alias{infer_haplotypes}
\title{EM-Based Inference of Multilocus HLA Haplotypes}
\usage{
infer_haplotypes(df, max_iter = 100, epsilon = 1e-05)
}
\arguments{
\item{df}{A data frame of unphased HLA genotypes. Column names must follow \if{html}{\out{<locus>}}_1 and \if{html}{\out{<locus>}}_2 structure.}

\item{max_iter}{Maximum number of EM iterations. Default is 100.}

\item{epsilon}{Convergence threshold for EM algorithm. Default is 1e-5.}
}
\value{
A list with:
\describe{
\item{frequencies}{A tibble of haplotypes and their estimated frequencies}
\item{posteriors}{A list of posterior diplotype probabilities per individual}
\item{top_diplotypes}{A tibble with one inferred diplotype per individual}
}
}
\description{
Runs EM algorithm on HLA genotype data to estimate haplotype frequencies
and assign most likely diplotype per individual.
}
