% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Plot_HLA_allele_count}
\alias{Plot_HLA_allele_count}
\title{Plot Count of Unique Alleles per HLA Gene}
\usage{
Plot_HLA_allele_count(hped)
}
\arguments{
\item{hped}{A data frame or tibble with HLA allele columns.}
}
\description{
Visualizes the number of unique observed alleles for each gene.
}
