% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{Plot_HLA_allele_frequency}
\alias{Plot_HLA_allele_frequency}
\title{Plot HLA Allele Frequencies by Gene}
\usage{
Plot_HLA_allele_frequency(hped, minFreq = 0.05)

Plot_HLA_allele_frequency(hped, minFreq = 0.05)
}
\arguments{
\item{hped}{A data frame or tibble with HLA allele columns.}

\item{minFreq}{Minimum frequency threshold (default = 0.05) above which allele labels are shown.}
}
\description{
Visualizes allele frequencies across genes, grouping rare alleles into "others".
Frequencies are computed using \code{calculate_HLA_frequency()}.

Visualizes allele frequencies across genes, grouping rare alleles into "others".
Frequencies are computed using \code{calculate_HLA_frequency()}.
}
