% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{reformat_typing_data}
\alias{reformat_typing_data}
\title{Reformat and Verbosely Validate Raw HLA Typing Data}
\usage{
reformat_typing_data(df, isfamilydata = TRUE)
}
\arguments{
\item{df}{A data frame or tibble containing raw HLA typing data.}

\item{isfamilydata}{Logical. Default TRUE. If FALSE, Family_Member sorting is skipped.}
}
\value{
A tibble with cleaned and reported HLA allele fields.
}
\description{
Cleans and formats HLA typing data by standardizing allele notations,
filling missing second allele calls, and reordering records.
}
